chore: move NoTableState to States folder
